+++ presentation/src/main/java/com/anytypeio/anytype/presentation/editor/editor/pattern/Pattern.kt
@@ -1,5 +1,5 @@
 package com.anytypeio.anytype.presentation.editor.editor.pattern
 
 enum class Pattern {
-    NUMBERED, DIVIDER, CHECKBOX, H1, H2, H3, QUOTE, TOGGLE, BULLET
+    NUMBERED, DIVIDER_LINE, DIVIDER_DOTS, CHECKBOX, H1, H2, H3, QUOTE, TOGGLE, BULLET, SNIPPET
 }